move conf in access
