+++ aias_common/access/configuration.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel, Field
+
+
+class StorageSettings(BaseModel, extra='allow'):
+    type: str = Field(title='Type of storage used')
+
+
+class AccessManagerSettings(BaseModel):
+    storages: list[StorageSettings] = Field(title="List of configurations for the available storages")
+    tmp_dir: str = Field(title="Temporary directory in which write files that will be deleted")
