+++ aias_common/access/logger.py
@@ -1,0 +1,9 @@
+import logging
+import os
+
+from aias_common.logger import CustomLogger
+
+
+class Logger(CustomLogger):
+    logger_name = "access"
+    level = os.getenv("ACCESS_LOGGER_LEVEL", logging.INFO)
