+++ aproc/core/settings.py
@@ -1,6 +1,7 @@
 from envyaml import EnvYAML
 from pydantic import BaseModel, Field
 
+from aias_common.access.storages.configuration import AccessManagerSettings
 from aproc.core.logger import Logger
 
 LOGGER = Logger.logger
@@ -10,15 +11,6 @@ class ProcessSettings(BaseModel, extra='allow'):
     name: str | None = Field(title="Name of the process")
     class_name: str | None = Field(title="Name of the process class")
     configuration: dict | None = Field(title="Process configuration")
-
-
-class StorageSettings(BaseModel, extra='allow'):
-    type: str = Field(title='Type of storage used')
-
-
-class AccessManagerSettings(BaseModel):
-    storages: list[StorageSettings] = Field(title="List of configurations for the available storages")
-    tmp_dir: str = Field(title="Temporary directory in which write files that will be deleted")
 
 
 class Settings(BaseModel, extra='allow'):
